rename: Response 객체를 dto 디렉터리 내부로 이동 (#301)
