+++ backend/src/test/java/com/celuveat/acceptance/restaurant/RestaurantLikeAcceptanceTest.java
@@ -23,7 +23,7 @@ import static com.celuveat.restaurant.fixture.RestaurantFixture.음식점;
 import com.celuveat.acceptance.common.AcceptanceTest;
 import com.celuveat.auth.application.OauthService;
 import com.celuveat.auth.domain.OauthMember;
-import com.celuveat.auth.presentation.SessionResponse;
+import com.celuveat.auth.presentation.dto.SessionResponse;
 import com.celuveat.common.SeedData;
 import com.celuveat.restaurant.application.dto.RestaurantLikeQueryResponse;
 import com.celuveat.restaurant.application.dto.RestaurantQueryResponse;
